requirements.txt
# ═══════════════════════════════════════════════════════
#  JARVIS V2/V3 — Python Dependencies
#  Install with: pip install -r requirements.txt
# ═══════════════════════════════════════════════════════

google-genai>=1.0.0
requests>=2.31.0
psutil>=6.1.0

# Mechanical/Vision Automation
pywin32>=311
pyautogui>=0.9.54
uiautomation>=2.0.29
numpy>=2.2.6
opencv-python>=4.12.0.88
pillow>=11.3.0

